Modify Interface & Implementation naming and add .gitignore
