--- src/main/java/com/google/flourbot/entity/EntityInterface.java
@@ -1,12 +1,0 @@
-package com.google.flourbot.entity;
-
-import java.lang.Exception;
-import java.util.Optional;
-import com.google.flourbot.datastorage.FirebaseDataStorage;
-
-import java.lang.InterruptedException;
-import java.util.concurrent.ExecutionException;
-
-public interface EntityInterface {
-    public Optional<Macro> getMacro(String userEmail, String macroName) throws InterruptedException, ExecutionException;
-}
