+++ src/main/java/com/google/flourbot/logic/Executor.java
@@ -1,16 +1,12 @@
 package com.google.flourbot.executor;
 
-import com.google.flourbot.datastorage.FirebaseDataStorage;
 import com.google.flourbot.entity.Macro;
-import com.google.flourbot.entity.trigger.Trigger;
 import com.google.flourbot.entity.action.Action;
 import com.google.flourbot.entity.action.ActionType;
-import com.google.flourbot.entity.EntityModule;
+import com.google.flourbot.entity.EntityModuleImplementation;
 
-import java.util.Map;
 import java.lang.Exception;
 import java.util.Optional;
-import com.google.cloud.firestore.QueryDocumentSnapshot;
 
 import java.lang.InterruptedException;
 import java.util.concurrent.ExecutionException;
@@ -18,18 +14,18 @@ import java.util.concurrent.ExecutionException;
 // The Logic class of the server
 public class Executor {
 
-    private final EntityModule entityModule;
+    private final EntityModuleImplementation entityModuleImplementation;
     //TODO: Add GoogleSheetHandler here when ready.
 
-    public Executor(EntityModule entityModule) throws Exception {
-        this.entityModule = EntityModule.getInstance();
+    public Executor(EntityModuleImplementation entityModuleImplementation) throws Exception {
+        this.entityModuleImplementation = EntityModuleImplementation.getInstance();
     }
 
     public String execute(String userEmail, String message) throws IllegalStateException, InterruptedException, ExecutionException {
         
         String macroName = message.split(" ")[0];
 
-        Optional<Macro> optionalMacro = entityModule.getMacro(userEmail, macroName);
+        Optional<Macro> optionalMacro = entityModuleImplementation.getMacro(userEmail, macroName);
         if (!optionalMacro.isPresent()) {
             throw new IllegalStateException("No macro named: " + macroName + " found!");
         }
